🖥️ ui: refactor SpaceContext hooks and update Space page components
