--- apps/web/src/ui/space/contexts/SpaceContext/hooks/index.ts
@@ -1,2 +1,0 @@
-export { useSpaceContext } from './useSpaceContext'
-export { useSpaceProvider } from './useSpaceProvider'
